requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-multipart==0.0.9
PyYAML==6.0.2
Pillow==10.4.0
typing-extensions>=4.8.0
